Fixes for device registration and persisting device status.
Closes #688
